Split util.py for v1 and common stuff
