+++ micro_sam/sam_annotator/_widgets.py
@@ -33,6 +33,7 @@ from qtpy.QtCore import QObject, Signal
 from superqt import QCollapsible
 
 from .. import util
+from ..v1.util import models
 from ..v1 import instance_segmentation
 from ..v1.multi_dimensional_segmentation import (
     PROJECTION_MODES,
@@ -450,7 +451,7 @@ class _WidgetBase(QtWidgets.QWidget):
         self.model_size = self._model_size_map[self._default_model_choice[4]]
 
         # Get all model options.
-        self.model_options = list(util.models().urls.keys())
+        self.model_options = list(models().urls.keys())
         # Filter out the decoders from the model list.
         self.model_options = [
             model
