+++ test/test_models/test_peft_sam.py
@@ -2,7 +2,7 @@ import unittest
 
 import torch
 
-import micro_sam.util as util
+from micro_sam.v1.util import get_sam_model
 
 
 class TestPEFTSam(unittest.TestCase):
@@ -20,49 +20,49 @@ class TestPEFTSam(unittest.TestCase):
     def test_lora_sam(self):
         from micro_sam.v1.models.peft_sam import PEFT_Sam, LoRASurgery
 
-        _, sam = util.get_sam_model(model_type=self.model_type, return_sam=True, device="cpu")
+        _, sam = get_sam_model(model_type=self.model_type, return_sam=True, device="cpu")
         peft_sam = PEFT_Sam(sam, rank=2, peft_module=LoRASurgery)
         self._check_output(peft_sam)
 
     def test_fact_sam(self):
         from micro_sam.v1.models.peft_sam import PEFT_Sam, FacTSurgery
 
-        _, sam = util.get_sam_model(model_type=self.model_type, return_sam=True, device="cpu")
+        _, sam = get_sam_model(model_type=self.model_type, return_sam=True, device="cpu")
         peft_sam = PEFT_Sam(sam, rank=2, peft_module=FacTSurgery)
         self._check_output(peft_sam)
 
     def test_attention_layer_peft_sam(self):
         from micro_sam.v1.models.peft_sam import PEFT_Sam, AttentionSurgery
 
-        _, sam = util.get_sam_model(model_type=self.model_type, return_sam=True, device="cpu")
+        _, sam = get_sam_model(model_type=self.model_type, return_sam=True, device="cpu")
         peft_sam = PEFT_Sam(sam, rank=2, peft_module=AttentionSurgery)
         self._check_output(peft_sam)
 
     def test_norm_layer_peft_sam(self):
         from micro_sam.v1.models.peft_sam import PEFT_Sam, LayerNormSurgery
 
-        _, sam = util.get_sam_model(model_type=self.model_type, return_sam=True, device="cpu")
+        _, sam = get_sam_model(model_type=self.model_type, return_sam=True, device="cpu")
         peft_sam = PEFT_Sam(sam, rank=2, peft_module=LayerNormSurgery)
         self._check_output(peft_sam)
 
     def test_bias_layer_peft_sam(self):
         from micro_sam.v1.models.peft_sam import PEFT_Sam, BiasSurgery
 
-        _, sam = util.get_sam_model(model_type=self.model_type, return_sam=True, device="cpu")
+        _, sam = get_sam_model(model_type=self.model_type, return_sam=True, device="cpu")
         peft_sam = PEFT_Sam(sam, rank=2, peft_module=BiasSurgery)
         self._check_output(peft_sam)
 
     def test_ssf_peft_sam(self):
         from micro_sam.v1.models.peft_sam import PEFT_Sam, SSFSurgery
 
-        _, sam = util.get_sam_model(model_type=self.model_type, return_sam=True, device="cpu")
+        _, sam = get_sam_model(model_type=self.model_type, return_sam=True, device="cpu")
         peft_sam = PEFT_Sam(sam, rank=2, peft_module=SSFSurgery)
         self._check_output(peft_sam)
 
     def test_adaptformer_peft_sam(self):
         from micro_sam.v1.models.peft_sam import PEFT_Sam, AdaptFormer
 
-        _, sam = util.get_sam_model(model_type=self.model_type, return_sam=True, device="cpu")
+        _, sam = get_sam_model(model_type=self.model_type, return_sam=True, device="cpu")
         peft_sam = PEFT_Sam(sam, rank=2, peft_module=AdaptFormer, projection_size=64, alpha=2.0, dropout=0.5)
         self._check_output(peft_sam)
